backend/app/services/recipe_service.py
from typing import List

import pinecone
from langchain.embeddings import OpenAIEmbeddings
from langchain.llms import OpenAI

from app.core.config import settings


class RecipeService:
    def __init__(self):
        # Initialize Pinecone
        pinecone.init(
            api_key=settings.PINECONE_API_KEY,
            environment=settings.PINECONE_ENVIRONMENT,
        )
        self.index = pinecone.Index(settings.PINECONE_INDEX_NAME)
        
        # Initialize OpenAI
        self.embeddings = OpenAIEmbeddings()
        self.llm = OpenAI()
    
    async def search_by_ingredients(self, ingredients: List[str]) -> List[dict]:
        """
        Search for recipes based on ingredients using RAG.
        """
        # TODO: Implement recipe search using Pinecone
        # For now, return mock data
        return []
    
    async def generate_recipe(self, ingredients: List[str]) -> dict:
        """
        Generate a new recipe based on ingredients using RAG.
        """
        # TODO: Implement recipe generation using OpenAI
        # For now, return mock data
        return {} 