backend/pyproject.toml
[tool.poetry]
name = "rag-recipe-backend"
version = "0.1.0"
description = "RAG-based recipe finder backend Lambda functions"
authors = ["David Ho <davidsw.ho02@gmail.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = ">=0.115.12,<0.116.0"
pinecone-client = ">=6.0.0,<7.0.0"
mangum = ">=0.19.0,<0.20.0"
pydantic = "^2.6.4"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
black = "^24.3.0"
isort = "^5.13.2"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
